practice-app/requirements.txt
psycopg2-binary
Django
djangorestframework
requests
django-dotenv
django-cors-headers